requirements.txt
# prod
Django==3.1.1
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
djangorestframework==3.12.2
django-storages[dropbox]==1.10.1

# dev and test
black==19.3b0 #20.8b0
flake8==3.8.3
isort==5.4.2